chore: reorganize/rename components
