+++ web/package.json
@@ -53,12 +53,9 @@
     "nookies": "^2.5.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.43.4",
-    "react-hooks-useform": "^0.1.86",
-    "react-select": "^5.7.0",
     "react-dropzone": "^14.2.3",
+    "react-select": "^5.7.0",
     "react-table": "^7.8.0",
-    "react-use-form-state": "^0.13.2",
     "sharp": "^0.31.3",
     "start-server-and-test": "^2.0.0",
     "styled-components": "^5.3.6",
